flows/pyproject.toml
[project]
name = "flows"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.10"
dependencies = [
    "apache-airflow==3.0.1",
    "apache-airflow-providers-common-compat==1.6.1",
    "apache-airflow-providers-common-sql==1.27.0",
    "apache-airflow-providers-docker==4.3.1",
    "apache-airflow-providers-postgres==6.1.3",
]

[dependency-groups]
dev = [
    "ipdb>=0.13.13",
]
test = [
    "mypy>=1.16.0",
    "ruff>=0.11.13",
]
